feat: Improve Access Key serialization in AddKey actions (#23)
* Improve Access Key serialization in AddKey actions

* Address review suggestions
