+++ Cargo.toml
@@ -19,6 +19,7 @@ itertools = "0.9.0"
 num-traits = "0.2.11"
 openssl-probe = { version = "0.1.2" }
 r2d2 = "0.8.8"
+serde = { version = "1", features = ["derive"] }
 serde_json = "1.0.55"
 tokio = { version = "0.2", features = ["sync", "time"] }
 tokio-diesel = "0.3.0"
